175s → 83s: use out parameter for hitRecord
This is one of the rare cases where the struct is copied so often.
And it is stack allocatable.
And it is so big.
And it goes through so many frames.

That an out parameter to skip a copy make sense.

---
Applied as is but will ... change in next commit