Improve performance by 5.5x : 143s to 26s (#26)
* 175s → 83s: use out parameter for hitRecord

This is one of the rare cases where the struct is copied so often.
And it is stack allocatable.
And it is so big.
And it goes through so many frames.

That an out parameter to skip a copy make sense.

---
Applied as is but will ... change in next commit

* My version of the same, and using original/current vec3. also updated the tests

* linter

* clearer and safer with the boolean

* Give up on go generics for typesafety as it impairs performance, inspired by @Jorropo's commit

* linter

* Add perf note in readme

* Add link to benchmark code

---------

Co-authored-by: Jorropo <jorropo.pgm@gmail.com>